Fixed a bug with automatic layer update
Filters that created (and consequently selected) new layers with
different rendering properties from the starting one (like sampling,
that creates a new layer rendered as point cloud) would trigger a layer
update event that would enable on the old layer the rendering passes
according to the GUI state dictated by the newly created one. Update
events are now handled according to the state of the layer object rather
than that of the GUI button.