Renaming basic layer creation function
trying also to clean up the mechanism of the data that is present on
the mesh. Corrected a few name exporting bug in the cpp to js macro.

Renamed MLJ.core.Scene.createCppMeshFile  -> MLJ.core.Scene.createLayer

Added addPerVertexNormal to all functions that create a layer that is a
point cloud and they generate a normals too.