Added Selection tool per vertexes or per faces
Changed "css/style.css": defined shape and design of selection square;
Changed "index.html" : included two new javascript files about selection
(Selection.js and Tool.js);
Changed "Manager.js": added _tool variable to manage the new Tool panel;
Added "Tool.js" a file used to build the selection widget with a set of
properties;
Changed "Scene.js": added getContrlos function to get the trackball
controls everywhere into the program
Changed "TabbedPane": the code added is used to allow the visualization
of the new Tool panel into the Toobar
Added "Selection.js": this file manage the first widget by which we are
able to select vertexes or faces on the current layer. Many tasks
include: set and install the new widget, manage the events related to
the selection region and send all needed parameters to c++ function to
initialize a camera and select object properly on the current mesh;
Added "ShotSelection.cpp" and "SimpleCamera.cpp": the first one is a new
Source c++ file that is able to manage the selection, it receive the
parameters from javascript and built a new camera object; the second one
contains a set of methods used to project a vertex from object to screen
coordinates and to check if a specific vertex is contained into a
specific selected box - selected region.