Added label maker
Defined in Scene.js a new function named makeTextLabel that allow to
make a new label in a given world point of the scene. The parameters are
the message to show up, the point where the message has to be showed up,
a set of parameters by which you can change the name of the label, the
font size, the font weight and so on and finally the THREE object in
which the label will be added.