Integration of the quality contour with the quality histogram
Added the option to enable the quality contour with the histogram.
Added a quality bar with the histogram that displays the values of the
quality range using a fixed number of bins and also displays the quality
contour lines if they are enabled.
Fixed minor bugs with the parameters of the histogram's GUI (before some
non valid parameters could be used, causing errors).