Added wedge split into vertex and tests
Added function to split wedge coordinates into vertex texture
coordinates
Added some testing on the Axis.js file, when the axis button is pressed
a new mesh will be shown with the applied texture. Apparently using a
custome texture cause the mesh to be completely black, while switching
to a simpler geometry cause the texture to be shown.
This is just texting and will be reverted back as soon as the texture
applying process will be complete and working