Added texture preloading
Texture are now loaded correctly before adding the overlay layer.
But if more meshes are loaded, the textures are not displayed correctly.
Every mesh will use the texture of the first loaded mesh, even when not
present, in that case every mesh will be black then enabling the
texturing.

If the first mesh has a texture, every following mesh will use the
texture no matter what, I still have to figure out what is causing the
issue