Fixed texture dynamic loading
Texture are now loaded dynamically from the emscripten's file system
the c++ function getTextureImage() will return an array with all the
images infos. The image will be then rebuilded in javascript according
to image's size and components and then used for creating a DataTexture

There is still some work to do:
When a mesh without texture is loaded BEFORE a mesh WITH a texture, the
one with the texture will stay black.
Instead, when a mesh with texture is loaded BEFORE a mesh WITHOUT a
texture, the latter will be always displayed with the last loaded
texture even though it does not have a texture of its own.